refactor: package 이동 및 main함수 수정(피드백 사항)
