+++ src/main/java/ladder/domain/GameResult.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package ladder.domain;
 
 import java.util.stream.IntStream;
