+++ src/main/java/ladder/domain/Line.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package ladder.domain;
 
 import java.util.ArrayList;
 import java.util.List;
